backend/requirements.txt
pymongo
python-dotenv
pyjwt
werkzeug
unittest 
mock